Don’t reload the component on every URL change
Currently in componentWillReceiveProps, the componentData is
reloaded if the URL changes no matter what. This is to prevent
a race condition if loadComponent returns a promise that resolves
after an URL changes that should load a different component.

Only reload the component in that particular race condition case.
Otherwise it causes a remount upon every URL change (it shouldn’t
when query parameters are added, for example).